* expert/student imitation learning example and visualization     * Save expert data under `imitation-data/{env}_{algo}_expert.tau`     * Imitate with reinforcement learning using cross entropy     * Save student trace under save random seed under `imitation-data/{env}_{algo}_student.tau` * TEST remove test-data folder after * MAKE add clean script
Took 20 hours 57 minutes

+++ cartpole-v0/helpers.py
@@ -2,13 +2,15 @@ import os
 import pickle
 
 
-def save(path, data):
+def save(path, data, no_overwrite=False):
     directory, fn = os.path.split(path)
     # make a new path
     try:
         os.makedirs(directory)
     except:
         pass
+    if no_overwrite and os.path.isfile(path):
+        raise Exception('data file already exists')
     with open(path, 'wb+') as f:
         pickle.dump(data, f)
 
@@ -18,7 +20,6 @@ def load(path):
     with open(path, 'rb') as f:
         data = pickle.load(f)
         return data
-
 
 # from ramuel.yaml import YAML
 #
